refactor: relocate FuturesStream tests to futures_stream file
