+++ src/utils/futures_stream.rs
@@ -84,3 +84,169 @@ impl<F: Future> Stream for FuturesStream<F> {
         Poll::Ready(Some(result))
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{
+        error::SubstreamError, mock::substream::MockSubstream, utils::futures_stream::FuturesStream,
+    };
+    use bytes::BytesMut;
+    use futures::StreamExt;
+
+    async fn get_data_from_substream(
+        mut substream: MockSubstream,
+    ) -> (Result<BytesMut, SubstreamError>, MockSubstream) {
+        let request = match substream.next().await {
+            Some(Ok(request)) => Ok(request),
+            Some(Err(error)) => Err(error),
+            None => Err(SubstreamError::ConnectionClosed),
+        };
+
+        (request, substream)
+    }
+
+    #[test]
+    fn add_substream() {
+        let mut set = FuturesStream::new();
+
+        let substream = MockSubstream::new();
+        set.push(get_data_from_substream(substream));
+
+        let substream = MockSubstream::new();
+        set.push(get_data_from_substream(substream));
+    }
+
+    #[tokio::test]
+    async fn poll_data_from_substream() {
+        let mut set = FuturesStream::new();
+
+        let mut substream = MockSubstream::new();
+        substream
+            .expect_poll_next()
+            .times(1)
+            .return_once(|_| Poll::Ready(Some(Ok(BytesMut::from(&b"hello"[..])))));
+        substream
+            .expect_poll_next()
+            .times(1)
+            .return_once(|_| Poll::Ready(Some(Ok(BytesMut::from(&b"world"[..])))));
+        substream.expect_poll_next().returning(|_| Poll::Pending);
+
+        set.push(get_data_from_substream(substream));
+        let (value, substream) = set.next().await.unwrap();
+        assert_eq!(value.unwrap(), BytesMut::from(&b"hello"[..]));
+
+        set.push(get_data_from_substream(substream));
+        let (value, _substream) = set.next().await.unwrap();
+        assert_eq!(value.unwrap(), BytesMut::from(&b"world"[..]));
+
+        assert!(futures::poll!(set.next()).is_pending());
+    }
+
+    #[tokio::test]
+    async fn substream_closed() {
+        let mut set = FuturesStream::new();
+
+        let mut substream = MockSubstream::new();
+        substream
+            .expect_poll_next()
+            .times(1)
+            .return_once(|_| Poll::Ready(Some(Ok(BytesMut::from(&b"hello"[..])))));
+        substream.expect_poll_next().times(1).return_once(|_| Poll::Ready(None));
+        substream.expect_poll_next().returning(|_| Poll::Pending);
+
+        set.push(get_data_from_substream(substream));
+
+        let (value, substream) = set.next().await.unwrap();
+        assert_eq!(value.unwrap(), BytesMut::from(&b"hello"[..]));
+
+        set.push(get_data_from_substream(substream));
+        let (value, _substream) = set.next().await.unwrap();
+        assert_eq!(value, Err(SubstreamError::ConnectionClosed));
+    }
+
+    #[tokio::test]
+    async fn poll_data_from_two_substreams() {
+        let mut set = FuturesStream::new();
+
+        // prepare first substream
+        let mut substream1 = MockSubstream::new();
+        substream1
+            .expect_poll_next()
+            .times(1)
+            .return_once(|_| Poll::Ready(Some(Ok(BytesMut::from(&b"hello"[..])))));
+        substream1
+            .expect_poll_next()
+            .times(1)
+            .return_once(|_| Poll::Ready(Some(Ok(BytesMut::from(&b"world"[..])))));
+        substream1.expect_poll_next().returning(|_| Poll::Pending);
+        set.push(get_data_from_substream(substream1));
+
+        // prepare second substream
+        let mut substream2 = MockSubstream::new();
+        substream2
+            .expect_poll_next()
+            .times(1)
+            .return_once(|_| Poll::Ready(Some(Ok(BytesMut::from(&b"siip"[..])))));
+        substream2
+            .expect_poll_next()
+            .times(1)
+            .return_once(|_| Poll::Ready(Some(Ok(BytesMut::from(&b"huup"[..])))));
+        substream2.expect_poll_next().returning(|_| Poll::Pending);
+        set.push(get_data_from_substream(substream2));
+
+        let expected: Vec<Vec<BytesMut>> = vec![
+            vec![
+                BytesMut::from(&b"hello"[..]),
+                BytesMut::from(&b"world"[..]),
+                BytesMut::from(&b"siip"[..]),
+                BytesMut::from(&b"huup"[..]),
+            ],
+            vec![
+                BytesMut::from(&b"hello"[..]),
+                BytesMut::from(&b"siip"[..]),
+                BytesMut::from(&b"world"[..]),
+                BytesMut::from(&b"huup"[..]),
+            ],
+            vec![
+                BytesMut::from(&b"siip"[..]),
+                BytesMut::from(&b"huup"[..]),
+                BytesMut::from(&b"hello"[..]),
+                BytesMut::from(&b"world"[..]),
+            ],
+            vec![
+                BytesMut::from(&b"hello"[..]),
+                BytesMut::from(&b"siip"[..]),
+                BytesMut::from(&b"huup"[..]),
+                BytesMut::from(&b"world"[..]),
+            ],
+        ];
+
+        // poll values
+        let mut values = Vec::new();
+
+        for _ in 0..4 {
+            let (value, substream) = set.next().await.unwrap();
+            values.push(value.unwrap());
+            set.push(get_data_from_substream(substream));
+        }
+
+        let mut correct_found = false;
+
+        for set in expected {
+            if values == set {
+                correct_found = true;
+                break;
+            }
+        }
+
+        if !correct_found {
+            panic!("invalid set generated");
+        }
+
+        // rest of the calls return `Poll::Pending`
+        for _ in 0..10 {
+            assert!(futures::poll!(set.next()).is_pending());
+        }
+    }
+}
